chore: split examples into individual files with runner, update README and polish
