+++ packages/inputlayer-py/examples/langchain/__init__.py
@@ -1,0 +1,1 @@
+"""LangChain integration examples for InputLayer."""
